feat(global-dashboard-widgets): Phase 3 — extract shared portfolio pieces
Move the reusable portfolio helpers out of the soon-to-be-deleted
PortfolioPage/roleLayout/rag.ts into a survivor-safe shared module
features/widgets/definitions/shared/portfolioMetrics.ts (G12 — the
sequencing gate before Phase 4 widgets + Phase 10 retirement).

Moved verbatim (no behavior change):
- CPI/SPI/VAC/TCPI MetricMetadata constants (from PortfolioPage).
- rag helpers: RagBand, RED_BAND_THRESHOLD, indexBand, ragBand, ragToStatus
  (from rag.ts — now deleted).
- cpiCostDistress (from roleLayout.ts).

Repointed all 4 importers (PortfolioPage, roleLayout, roleLayout.test,
FilterBar) to the shared module. LEAD_METADATA/LeadMetric stay in the
doomed files (they die in Phase 10). Migrated tests: rag.test.ts (17) +
the cpiCostDistress block from roleLayout.test (5) -> portfolioMetrics.test
(22). Deleted rag.ts + rag.test.ts.

Gates: tsc clean (only the 2 known pre-existing useStreamingChat errors);
lint clean; 180 portfolio+widget tests pass (incl. 22 migrated +
PortfolioPage.role/distressPagination regression green; roleLayout.test
now 9, was 14).

Refs: design doc §7 Phase 3; gap G12.
